Split MCP bounty attempt tests
